refactor: change job names and create missing
