+++ decidim-bulletin_board-app/app/jobs/end_vote_job.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+# A job to process end vote messages
+class EndVoteJob < ApplicationJob
+  include ProcessEnqueuedMessage
+
+  command_class EndVote
+  queue_as :vote
+end
