+++ decidim-bulletin_board-app/app/jobs/start_vote_job.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+# A job to process start vote messages
+class StartVoteJob < ApplicationJob
+  include ProcessEnqueuedMessage
+
+  command_class StartVote
+  queue_as :vote
+end
